Split the four oversized MIR files at family boundaries
